Changed productDetail and AddToCart button
